Big changes (detail in full description).
This commit remove ClientCommandController and ServerController.
CommandController don't do the getting input anymore. On client side, this
is done with client.repl.input.UserInputProvider and on server side,
it is easily extract input from the query.

Also everything that was considered to be `event` is now changed into callback
for single time use.